Refactor: Menu를 CalculatorDriver로부터 분리
